Adds signature helpers, updates wasm

+++ console/collections/src/merkle_tree/mod.rs
@@ -289,7 +289,7 @@ impl<E: Environment, LH: LeafHash<Hash = PH::Hash>, PH: PathHash<Hash = Field<E>
 #[inline]
 fn tree_depth<const DEPTH: u8>(tree_size: usize) -> Result<u8> {
     // Ensure the tree size is less than 2^52 (for casting to an f64).
-    let tree_depth = match tree_size < 4503599627370496 {
+    let tree_depth = match (tree_size as u64) < 4503599627370496_u64 {
         // Compute the log2 of the tree size.
         true => (tree_size as f64).log2(),
         false => bail!("Tree size must be less than 2^52"),
